+++ doc/source/conf.py
@@ -10,8 +10,10 @@
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
+
 import os
 import sys
+from datetime import date
 
 import sphinx_rtd_theme
 
@@ -21,7 +23,7 @@ sys.path.insert(0, os.path.abspath('../..'))
 # -- Project information -----------------------------------------------------
 
 project = 'tmtoolkit'
-copyright = '2020, Markus Konrad'
+copyright = f'{date.today().year}, Markus Konrad'
 author = 'Markus Konrad'